flask-jwt-auth/requirements.txt
Flask==2.2.5
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
Flask-Cors==4.0.0
bcrypt==4.1.2
langchain
langchain-community
langchain-openai
pgvector
